+++ tests/test-715byte_swapped_ac3.rb
@@ -1,0 +1,5 @@
+#!/usr/bin/ruby -w
+
+# T_715byte_swapped_ac3
+describe "mkvmerge / byte-swapped AC-3"
+test_merge "data/ac3/byte-swapped.ac3"